brda toga uradjeno nije provereno s gemini, stali kod "Administrator ima uvid u podatke o svim zaposlenim kao i prihode i rashode za određeni period." treba dodati parove <zaposlen, plata> i csv nista nisu dorbi podaci. Po negde dodati novi atributi u klasama.

+++ src/model/Zaposleni.java
@@ -40,7 +40,7 @@ public class Zaposleni extends Osoba {
 		return osnovnaPlata;
 	}
 
-	public void setOsnovnaPlata(int osnovnaPlata) {
+	public void setOsnovnaPlata(double osnovnaPlata) {
 		this.osnovnaPlata = osnovnaPlata;
 	}
